[STEP-3] Car 추가
