+++ racingcar/src/main/kotlin/racingcar/domain/car/Car.kt
@@ -1,0 +1,25 @@
+package racingcar.domain.car
+
+class Car(
+    private var id: Int = 0,
+    private var position: Int = 0,
+    private val accelerator: Accelerator = Accelerator(),
+) {
+    val currentPosition
+        get() = position
+
+    fun getId() = id
+
+    fun init(id: Int): Car {
+        if (this.id == 0) {
+            this.id = id
+        }
+        return this
+    }
+
+    fun tryForward() {
+        if (accelerator.tryForward()) {
+            position++
+        }
+    }
+}
